% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/these_boots_are_made_for.R
\name{bootstrap}
\alias{bootstrap}
\title{A self-written Bootstrap function}
\usage{
bootstrap(n = NULL, x = NULL, iter = NULL, mode = "mean", as_tbl = FALSE)
}
\arguments{
\item{x}{A vector of type nummeric (the sample data)}

\item{iter}{Number of bootstrapped samples to generate}

\item{mode}{Which statistic do you want to return - 
currently implemented are "mean", "median" and "sd"}
}
\description{
A self-written Bootstrap function
}
